🚚 Move components to model
