+++ anomalib/data/utils/download_progress_bar.py
