+++ anomalib/deploy/inferencers/torch.py
@@ -23,9 +23,9 @@ import torch
 from omegaconf import DictConfig, ListConfig
 from torch import Tensor
 
-from anomalib.components.anomaly_models import AnomalyModule
 from anomalib.deploy.optimize import get_model_metadata
 from anomalib.models import get_model
+from anomalib.models.components.anomaly_models import AnomalyModule
 from anomalib.pre_processing.pre_process import PreProcessor
 
 from .base import Inferencer
