+++ anomalib/models/components/anomaly_models/anomaly_module.py
