+++ anomalib/models/components/dimensionality_reduction/pca.py
@@ -19,7 +19,7 @@ from typing import Union
 import torch
 from torch import Tensor
 
-from anomalib.components.anomaly_models.dynamic_module import DynamicBufferModule
+from anomalib.models.components.anomaly_models.dynamic_module import DynamicBufferModule
 
 
 class PCA(DynamicBufferModule):
