+++ anomalib/models/components/feature_extractors/__init__.py
