+++ anomalib/models/components/feature_extractors/feature_extractor.py
