+++ anomalib/models/components/neighbors/nearest_neighbors.py
@@ -22,7 +22,7 @@ from typing import Tuple
 import torch
 from torch import Tensor
 
-from anomalib.components.anomaly_models.dynamic_module import DynamicBufferModule
+from anomalib.models.components.anomaly_models.dynamic_module import DynamicBufferModule
 
 
 class NearestNeighbors(DynamicBufferModule):
