+++ anomalib/models/dfkde/normality_model.py
@@ -20,8 +20,8 @@ from typing import Optional, Tuple
 import torch
 from torch import Tensor, nn
 
-from anomalib.components.dimensionality_reduction.pca import PCA
-from anomalib.components.stats.kde import GaussianKDE
+from anomalib.models.components.dimensionality_reduction.pca import PCA
+from anomalib.models.components.stats.kde import GaussianKDE
 
 
 class NormalityModel(nn.Module):
