+++ anomalib/models/padim/model.py
@@ -27,9 +27,11 @@ from kornia import gaussian_blur2d
 from omegaconf import DictConfig, ListConfig
 from torch import Tensor, nn
 
-from anomalib.components.anomaly_models import AnomalyModule
-from anomalib.components.feature_extractors.feature_extractor import FeatureExtractor
-from anomalib.components.stats.multi_variate_gaussian import MultiVariateGaussian
+from anomalib.models.components.anomaly_models import AnomalyModule
+from anomalib.models.components.feature_extractors.feature_extractor import (
+    FeatureExtractor,
+)
+from anomalib.models.components.stats.multi_variate_gaussian import MultiVariateGaussian
 from anomalib.pre_processing.tiler import Tiler
 
 __all__ = ["PadimLightning"]
