+++ anomalib/utils/callbacks/save_to_csv.py
@@ -5,7 +5,7 @@ import numpy as np
 import pandas as pd
 from pytorch_lightning import Callback, Trainer
 
-from anomalib.components.anomaly_models import AnomalyModule
+from anomalib.models.components.anomaly_models import AnomalyModule
 
 
 class SaveToCSVCallback(Callback):
