+++ tests/deploy/test_inferencer.py
@@ -23,11 +23,11 @@ import torch
 from omegaconf import DictConfig, ListConfig
 from pytorch_lightning import Trainer
 
-from anomalib.config import get_configurable_parameters
 from anomalib.data import get_datamodule
 from anomalib.deploy.inferencers import OpenVINOInferencer, TorchInferencer
 from anomalib.deploy.optimize import export_convert
 from anomalib.models import get_model
+from anomalib.utils.config import get_configurable_parameters
 from tests.helpers.dataset import TestDataset, get_dataset_path
 from tests.helpers.inference import MockImageLoader, get_meta_data
