+++ tools/test.py
@@ -18,10 +18,10 @@ from argparse import ArgumentParser, Namespace
 
 from pytorch_lightning import Trainer
 
-from anomalib.config import get_configurable_parameters
-from anomalib.core.callbacks import get_callbacks
 from anomalib.data import get_datamodule
 from anomalib.models import get_model
+from anomalib.utils.callbacks import get_callbacks
+from anomalib.utils.config import get_configurable_parameters
 
 
 def get_args() -> Namespace:
