+++ tools/train.py
@@ -23,11 +23,11 @@ from argparse import ArgumentParser, Namespace
 
 from pytorch_lightning import Trainer, seed_everything
 
-from anomalib.config import get_configurable_parameters
-from anomalib.core.callbacks import get_callbacks
 from anomalib.data import get_datamodule
-from anomalib.loggers import get_logger
 from anomalib.models import get_model
+from anomalib.utils.callbacks import get_callbacks
+from anomalib.utils.config import get_configurable_parameters
+from anomalib.utils.loggers import get_logger
 
 
 def get_args() -> Namespace:
